move several types from fwdpp/sugar into fwdpp
