+++ examples/K_linked_regions_multilocus.cc
@@ -13,7 +13,7 @@
 #include <sstream>
 #include <fwdpp/debug.hpp>
 // Use mutation model from sugar layer
-#include <fwdpp/sugar/popgenmut.hpp>
+#include <fwdpp/popgenmut.hpp>
 #include <fwdpp/sampling_functions.hpp>
 using mtype = fwdpp::popgenmut;
 #define MULTILOCUS_SIM
