+++ examples/common_ind.hpp
@@ -16,15 +16,15 @@
    mutation type
 */
 #ifdef SINGLEPOP_SIM
-#include <fwdpp/sugar/slocuspop.hpp>
+#include <fwdpp/slocuspop.hpp>
 using singlepop_t = fwdpp::slocuspop<mtype>;
 #elif defined(MULTILOCUS_SIM)
-#include <fwdpp/sugar/mlocuspop.hpp>
+#include <fwdpp/mlocuspop.hpp>
 using multiloc_t = fwdpp::mlocuspop<mtype>;
 #endif
 
 // RNG type
-#include <fwdpp/sugar/GSLrng_t.hpp>
+#include <fwdpp/GSLrng_t.hpp>
 
 using GSLrng = fwdpp::GSLrng_t<fwdpp::GSL_RNG_MT19937>;
