+++ fwdpp/poptypes/Makefile.am
@@ -1,3 +1,3 @@
-pkgincludedir=$(prefix)/include/fwdpp/sugar/poptypes
+pkgincludedir=$(prefix)/include/fwdpp/poptypes
 
 pkginclude_HEADERS=tags.hpp popbase.hpp slocuspop.hpp mlocuspop.hpp
