+++ fwdpp/sugar/Makefile.am
@@ -1,5 +1,5 @@
 pkgincludedir=$(prefix)/include/fwdpp/sugar
 
-SUBDIRS=gsl poptypes
+SUBDIRS=gsl 
 
-pkginclude_HEADERS=sugar.hpp slocuspop.hpp mlocuspop.hpp GSLrng_t.hpp popgenmut.hpp add_mutation.hpp change_neutral.hpp
+pkginclude_HEADERS=sugar.hpp add_mutation.hpp change_neutral.hpp
