+++ fwdpp/sugar/sugar.hpp
@@ -5,9 +5,7 @@
 #ifndef __FWDPP_SUGAR_SUGAR_HPP__
 #define __FWDPP_SUGAR_SUGAR_HPP__
 
-#include <fwdpp/sugar/GSLrng_t.hpp>
-#include <fwdpp/sugar/multiloc.hpp>
 #include <fwdpp/sugar/popgenmut.hpp>
-#include <fwdpp/sugar/singlepop.hpp>
 #include <fwdpp/sugar/add_mutation.hpp>
+
 #endif
