+++ testsuite/integration/sugar_multilocusTest.cc
@@ -11,8 +11,8 @@
 #include <numeric>
 #include <boost/test/unit_test.hpp>
 #include <fwdpp/diploid.hh>
-#include <fwdpp/sugar/GSLrng_t.hpp>
-#include <fwdpp/sugar/mlocuspop.hpp>
+#include <fwdpp/GSLrng_t.hpp>
+#include <fwdpp/mlocuspop.hpp>
 #include <fwdpp/forward_types_serialization.hpp>
 #include <fwdpp/io/serialize_population.hpp>
 #include "../fixtures/sugar_fixtures.hpp"
