+++ testsuite/unit/test_mutate_recombine.cc
@@ -1,8 +1,8 @@
 #include <iostream>
 #include <cassert>
 #include <boost/test/unit_test.hpp>
-#include <fwdpp/sugar/slocuspop.hpp>
-#include <fwdpp/sugar/popgenmut.hpp>
+#include <fwdpp/slocuspop.hpp>
+#include <fwdpp/popgenmut.hpp>
 #include <fwdpp/mutate_recombine.hpp>
 #include <testsuite/fixtures/sugar_fixtures.hpp>
